problems/src/test/java/_1051_1100/_1079_Letter_Tile_Possibilities/SolutionITest.java
package _1051_1100._1079_Letter_Tile_Possibilities;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author zhixi
 * @version 1.0 on 11/29/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final String tiles = "AAB";
    assertEquals(8, solution.numTilePossibilities(tiles));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase2(Solution solution) {
    final String tiles = "AAABBC";
    assertEquals(188, solution.numTilePossibilities(tiles));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase3(Solution solution) {
    final String tiles = "V";
    assertEquals(1, solution.numTilePossibilities(tiles));
  }

  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}